feat: implement customisation of Explorers and Strategies configs (#22)
* feat: implement customisation of Explorers and Strategies configs

* feat: add java docs

+++ src/main/java/org/shadow/config/model/OrderConfiguration.java
@@ -1,8 +1,6 @@
 package org.shadow.config.model;
 
-import java.math.BigDecimal;
-
 public record OrderConfiguration(
-    BigDecimal allowedOrderPercentageFromDeposit,
-    Integer allowedOrderFuturesMultiplier,
-    BigDecimal stopLossRequiredPercentage) {}
+    double allowedOrderPercentageFromDeposit,
+    int allowedOrderFuturesMultiplier,
+    double stopLossRequiredPercentage) {}
